SearchBar Update
Routing Fixed
